#392 chore : Book 패키지 테스트 코드 패키지 정리
